+++ packages/core/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.0.20
+
+## Checkbox
+- Removed `isRequired` prop type specification on `name` & `value`
+
+## TextArea
+- Removed `isRequired` prop type specification on `onFocus`, `onBlur`, `onKeyDown`, & `onChange`
+- Changed default props for `defaultValue` & `value` from `null` to `undefined`
+  - Otherwise, React triggers this warning: "Warning: [your component here] contains a textarea with both value and defaultValue props. Textarea elements must be either controlled or uncontrolled (specify either the value prop, or the defaultValue prop, but not both). Decide between using a controlled or uncontrolled textarea and remove one of these props. More info: https://fb.me/react-controlled-components"
+
 # 0.0.19
 
 ## DropDown
@@ -9,7 +19,6 @@
 - Added `required` prop to TextArea Form Control
 - Added stories to TextArea stories to reflect `error` prop
 - Added story to TextArea stories to reflect `required` prop
-
 
 # 0.0.18